requirements.txt
openai==1.2.0
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.28.1
tqdm==4.64.1
pytesseract==0.3.10
pdf2image==1.16.0
pillow==11.1.0
numpy==2.2.3